[common/outb,host/outb,guest] moved OutBAction to common lib
OutBAction was duplicated across the host and guest.
