+++ src/hyperlight_common/src/lib.rs
@@ -36,3 +36,6 @@ pub mod flatbuffer_wrappers;
 mod flatbuffers;
 /// cbindgen:ignore
 pub mod mem;
+
+/// cbindgen:ignore
+pub mod outb;
