+++ src/hyperlight_guest/src/lib.rs
@@ -24,8 +24,9 @@ use buddy_system_allocator::LockedHeap;
 use guest_function_register::GuestFunctionRegister;
 use hyperlight_common::flatbuffer_wrappers::guest_error::ErrorCode;
 use hyperlight_common::mem::{HyperlightPEB, RunMode};
+use hyperlight_common::outb::OutBAction;
 
-use crate::host_function_call::{outb, OutBAction};
+use crate::host_function_call::outb;
 extern crate alloc;
 
 // Modules
